Basketball Rebuild in progress!
All of the previous basketball scripts will be labeled old_*. The
rebuilt basketball scripts will generally use the same names.

ArcadeScene.unity
- Distinction: The old basketball game is located in
Level/Canvas/BasketballPanel. The new basketball game is located in
Level/Background/Basketball Camera (name to be changed)
- IMPORTANT CHANGE! The Canvas has its Canvas Scaler component updated
to have its screen match mode to be expand.

ArcadeDialogue.yarn
- The prizebooth dialogue now no longer talks about having tickets.
