+++ Assets/Scripts/Levels/Level6Arcade/Basketball Cam Render Texture.renderTexture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1459eca14ded5804c9801837bfb291e9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
