+++ Assets/Scripts/Levels/Level6Arcade/BasketballNet.cs
@@ -4,19 +4,16 @@ using UnityEngine;
 
 public class BasketballNet : MonoBehaviour
 {
-	// Attached Unity Objects
-
-	// Fields
-	public bool trackMouse = false;
-	public static readonly float LEFT_LIMIT = -440f;
-	public static readonly float RIGHT_LIMIT = LEFT_LIMIT * -1f;
-
 	void Update()
 	{
-		// update the actual position using the mouse position
-		transform.position = new Vector3(Input.mousePosition.x, transform.position.y, transform.position.z);
-		// calculate an updated x position using the transform's local position, and update localPosition
-		float xPos = Mathf.Clamp(transform.localPosition.x, LEFT_LIMIT, RIGHT_LIMIT);
-		transform.localPosition = new Vector3(trackMouse?xPos:0f, transform.localPosition.y, transform.localPosition.z);
+		float mousePercent = Input.mousePosition.x / Screen.width;
+
+		// uses seemingly random number -6 -> 6. For a better fix, grab the sprite width. However, it seems unnecessary
+		float xPos = Mathf.Clamp(-8 + (16 * mousePercent), -6, 6);
+
+
+		transform.localPosition = new Vector3(xPos, transform.localPosition.y, transform.localPosition.z);
+
+		// Debug.Log(xPos);
 	}
 }
