+++ Assets/Scripts/Levels/Level6Arcade/BasketballNet.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b867de93b9d398f4482aaa15e52cb2cd
+guid: 49b4345bd22bb764fb6bf1b801c8a045
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
