+++ Assets/Scripts/Levels/Level6Arcade/Yarn/ArcadeDialogue.yarn
@@ -145,6 +145,6 @@ title: PrizeBooth
 tags:
 ---
 <<ChangeSpeaker Player>>
-    This is the prize booth. Too bad I don't have any tickets.
+    This is the prize booth. I don't see anything of interest here.
 <<exit>>
 ===
